added results to editpayroll, moved saving to history to download email
