+++ React/src/pages/EditPayroll/EditPayroll.module.css
@@ -24,7 +24,7 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    width: 80%;
+    width: 95%;
     padding: 20px;
     margin-bottom: 30px;
     background-color: white;
@@ -59,8 +59,11 @@
     background-color: #065999;
 }
 
-.formGroup {
+.resultSection {
+    width: 40%;
     display: flex;
-    align-items: center;
-    padding-bottom: 20px;
+    flex-direction: column;
+    margin-top: 10px;
+    margin-right: 40px;
+    text-align: left;
 }
