+++ React/src/pages/_resultsInfo/ResultsInfo.module.css
@@ -1,0 +1,27 @@
+.formSection {
+    width: 80%;
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 20px;
+}
+
+.formGroup {
+    display: flex;
+    align-items: center;
+    padding-bottom: 10px;
+}
+
+.formGroup label {
+    width: 100px;
+    margin-right: 20px;
+}
+
+.formGroup input {
+    margin-bottom: 10px;
+    pointer-events:none;
+    border: none;
+    flex: 1;
+    padding: 6px;
+    background-color: #c7dae8;
+    border-radius: 4px;
+}
